+++ changelog.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- `revoke_userid` method for keystore
+- `add_userid` method for keystore
+- `update_password` method for keystore
 - ECDH decryption on smartcard for Curve25519 only
 - Adds `decrypt_file_on_card` function
 - Adds `decrypt_bytes_on_card` function